% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Brasileiro_A.R
\name{Brasileiro_A}
\alias{Brasileiro_A}
\alias{Brasileiro_A.}
\title{Function for data of Campeonato Brassileiro .}
\usage{
Brasileiro_A(server = NULL, type = NULL, year = NULL, ...)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class "SofaScore".}

\item{type}{esse objeto descreve se a tabla é geral, para os jogos em casa ou jogos fora de casa.}

\item{year}{esse objeto informa o ano que se deseja a tabela do campeonato.}

\item{...}{argumentos adicionais passados para outros métodos.}
}
\value{
Abre a página do sofa score.
}
\description{
Function for data of Campeonato Brassileiro .
}
