% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Close_Sofa.R
\name{Close_Sofa}
\alias{Close_Sofa}
\alias{Close_Sofa.}
\title{Close the page .}
\usage{
Close_Sofa(server = NULL, ...)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class "SofaScore".}

\item{...}{argumentos adicionais passados para outros métodos.}
}
\value{
Close the sofa score web page.
}
\description{
Close the page .
}
