% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapa_Calor.R
\name{Mapa_Calor}
\alias{Mapa_Calor}
\alias{Mapa_Calor.}
\title{Função que retorna o mapa de calor do jogador.}
\usage{
Mapa_Calor(server = NULL, player = NULL, team = NULL)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class remoteDriver}

\item{player}{esse objeto descreve o jogador que deseja saber o mapa de calor em campo.}

\item{team}{esse objeto descreve o time que o jogador pertence atualmente.}
}
\value{
retorna o mapa de calor do jogador.
}
\description{
Função que retorna o mapa de calor do jogador.
}
