% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Open_Sofa.R
\name{Open_Sofa}
\alias{Open_Sofa}
\alias{Open_Sofa.}
\title{Open the page .}
\usage{
Open_Sofa(browser = "firefox", ...)
}
\arguments{
\item{browser}{esse objeto é o nome do browser que o R deverá utilizar.}

\item{...}{argumentos adicionais passados da função "remoteDriver".}
}
\value{
Abre a página do sofa score.
}
\description{
Open the page .
}
