% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Players.R
\name{Players}
\alias{Players}
\alias{Players.}
\title{Função que retorna o elenco do time.}
\usage{
Players(server = NULL, team = NULL, position = NULL)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class remoteDriver}

\item{team}{esse objeto descreve o time que deseja saber o elenco.}

\item{position}{esse objeto informa a posição dos jogadores que deseja extrair informações.}
}
\value{
retorna elenco o time.
}
\description{
Função que retorna o elenco do time.
}
