% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sofa_score.R
\name{Sofa_score}
\alias{Sofa_score}
\alias{Sofa_score.}
\title{Open the page .}
\usage{
Sofa_score(port = NULL, chromever = NULL, ...)
}
\arguments{
\item{port}{esse objeto precisa ser um código da porta que você utiliza pra abrir o navegador.}

\item{chromever}{esse objeto é o código atualizado do crome que utiliza.}

\item{...}{argumentos adicionais passados para outros métodos.}
}
\value{
Abre a página do sofa score.
}
\description{
Open the page .
}
