% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statiscs_Games.R
\name{Statiscs_Games}
\alias{Statiscs_Games}
\alias{Statiscs_Games.}
\title{Função para trazer as estatísticas dos ultimos jogos do time.}
\usage{
Statiscs_Games(server = NULL, team = NULL, time = 1)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class "SofaScore".}

\item{team}{esse objeto descreve o time que deseja saber o elenco.}

\item{time}{tempo necessário para o sistema carregar páginas na web em segundos.}
}
\value{
retorna últimos jogos.
}
\description{
Função para trazer as estatísticas dos ultimos jogos do time.
}
