% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{TabLink}
\alias{TabLink}
\title{Tabel for links (TabLink) data set}
\format{
A data frame with 13 rows and 2 variables
}
\usage{
data(TabLink)
}
\description{
Data table indicating the team and the link to the Sofa Score page it has.
}
\references{
R and site of Sofa Score
}
\keyword{datasets}
