% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TabLinkPlayers.R
\name{TabLinkPlayers}
\alias{TabLinkPlayers}
\alias{TabLinkPlayers.}
\title{Função que retorna o elenco do time.}
\usage{
TabLinkPlayers(server = NULL, team = NULL)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class remoteDriver}

\item{team}{esse objeto descreve o time que deseja saber o elenco.}
}
\value{
retorna elenco o time.
}
\description{
Função que retorna o elenco do time.
}
