% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Last_Games.R
\name{Last_Games}
\alias{Last_Games}
\alias{Last_Games.}
\title{Função para trazer os ultimos jogos do time.}
\usage{
Last_Games(server = NULL, team = NULL)
}
\arguments{
\item{server}{connection with the SofaScore website by the Sofa_score() function, this object should be of the class "SofaScore".}

\item{team}{esse objeto descreve o time que deseja saber o elenco.}
}
\value{
retorna últimos jogos.
}
\description{
Função para trazer os ultimos jogos do time.
}
