% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sofa_score.R
\name{Sofa_score}
\alias{Sofa_score}
\alias{Sofa_score.}
\title{Open the page .}
\usage{
Sofa_score(browser = "firefox", ...)
}
\arguments{
\item{browser}{esse objeto é o nome do browser que o R deverá utilizar.}

\item{...}{argumentos adicionais passados da função "remoteDriver".}
}
\value{
Abre a página do sofa score.
}
\description{
Open the page .
}
